refactor: 중복확인 예외처리 리팩토링
