+++ src/main/java/baseball/Application.java
@@ -9,19 +9,6 @@ import java.util.Set;
 
 public class Application {
 
-    // 중복 검사 함수
-    public static boolean isDuplicationExist(String randomNum) {
-        ArrayList<Character> randomNumList = new ArrayList<Character>();
-        for (int i=0; i<3; i++){
-            randomNumList.add(randomNum.charAt(i));
-        }
-        Set<Character> randNumSet = new HashSet<>(randomNumList);
-        if (randNumSet.size()!=3) {
-            return true;
-        }
-        return false;
-    }
-
     public static boolean isCharacterExist(String randomNum) {
         try{
             Integer.parseInt(randomNum);
