+++ src/main/java/baseball/BaseballException.java
@@ -1,6 +1,22 @@
 package baseball;
 
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Set;
+
 public class BaseballException {
 
+    // 중복 검사 함수
+    public static boolean isDuplicationExist(String randomNum) {
+        ArrayList<Character> randomNumList = new ArrayList<Character>();
+        for (int i=0; i<3; i++){
+            randomNumList.add(randomNum.charAt(i));
+        }
+        Set<Character> randNumSet = new HashSet<>(randomNumList);
+        if (randNumSet.size()!=3) {
+            return true;
+        }
+        return false;
+    }
 
 }
